Add files via upload
In this challenge, we lay out buttons inside a div and modify their labels after each click event on one of the buttons. Check out the attached tutorial for learning materials.

Task 
We want to create nine buttons enclosed in a div, laid out so they form a  3*3 grid. Each button has a distinct label from 1 to 10 , and the labels on the outer buttons must rotate in the clockwise direction each time we click the middle button.